% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjust_sample_number.R
\name{adjust_sample_number}
\alias{adjust_sample_number}
\title{Adjust the sample size per selected unit}
\usage{
adjust_sample_number(
  data,
  budget,
  col_estimated,
  col_adjusted = "justert_ant_prover",
  adjust_by
)
}
\arguments{
\item{data}{Data frame}

\item{budget}{The total budgeted sample number.}

\item{col_estimated}{The name of the column with the sample size per unit that should be adjusted.}

\item{col_adjusted}{The name of the column with the adjusted sample size per unit.}

\item{adjust_by}{The maximum number of samples that one should adjust by.}
}
\value{
A data frame with a new column with an adjusted sample number.
}
\description{
Adds new column with an adjusted the sample size per selected
     unit. The total sample size is adjusted to be in accord with the total
     budgetted sample size.
}
\details{
The sample size should first be estimated by percentage or similar.
}
\author{
Petter Hopp Petter.Hopp@vetinst.no
}
