% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjust_samples_to_budget.R
\name{adjust_samples_to_budget}
\alias{adjust_samples_to_budget}
\title{Adjust the sample size per selected unit}
\usage{
adjust_samples_to_budget(
  data,
  group = NULL,
  budget,
  sample_to_adjust,
  adjusted_sample = "justert_ant_prover",
  adjust_by
)
}
\arguments{
\item{data}{Data frame}

\item{group}{\[character\]. Vector with group variables.}

\item{budget}{The total budgeted sample number or a column in data with the number of samples.}

\item{sample_to_adjust}{The name of the column with the sample number per unit that should be adjusted.}

\item{adjusted_sample}{The name of the column with the adjusted sample number per unit.}

\item{adjust_by}{The maximum number of samples that one should adjust by.}
}
\value{
A data frame with a new column with an adjusted sample number.
}
\description{
Adds new column with an adjusted the sample size per selected
     unit. The total sample size is adjusted to be in accord with the total
     budgeted sample size.
}
\details{
The sample size should first be estimated by percentage or similar.
}
\examples{
library(OKplan)
# Add data frame with sample number to adjust
x <- as.data.frame(cbind(c(1:10),
                         c(24, 30, 36, 12, 6, 18, 6, 0, 0, 0)))
colnames(x) <- c("id", "sample")

# Adjust total sample number to budget
x <- adjust_samples_to_budget(data = x,
                              budget = 150,
                              sample_to_adjust = "sample",
                              adjusted_sample = "new_sample",
                              adjust_by = 4)

# Adjust total sample number to budget per group
total_budget <- 60 # same budget for all groups

# Add data frame with sample number to adjust
x <- as.data.frame(cbind(c(1:10),
                         c(rep("x", 5), rep("y", 5)),
                         c(24, 18, 6, 0, 30, 36, 12, 6, 0, 0)))
colnames(x) <- c("id", "xy", "sample")

x2 <- adjust_samples_to_budget(data = x,
                               group = "xy",
                               budget = total_budget,
                               sample_to_adjust = "sample",
                               adjusted_sample = "new_sample",
                               adjust_by = 6)

}
\author{
Petter Hopp Petter.Hopp@vetinst.no
}
