% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/append_date_generated_line.R
\name{append_date_generated_line}
\alias{append_date_generated_line}
\title{Add new last row with the generated date}
\usage{
append_date_generated_line(
  data,
  pretext = "Datauttrekket er gjort",
  date = format(Sys.Date(), "\%d/\%m/\%Y")
)
}
\arguments{
\item{data}{[\code{data.frame}]\cr
The data to which an additional row with the generate date should be added.}

\item{pretext}{[\code{character(1)}]\cr
The explaining text before the date value. Defaults to "Datauttrekket ble gjort".}

\item{date}{[\code{character(1)}]\cr
Date for generating the data. Defaults to
\ifelse{html}{\code{\link[base]{Sys.Date}}}{\code{base::Sys.Date}}.}
}
\value{
\code{data.frame} with two additional rows, one empty and one with the generated date in the first column.
}
\description{
Function to add new row with the generated date. Before the date
    a pretext can be set.
}
\details{
Two rows are added to the data frame, the first is empty, the second
    has the generated date in the first column.
}
\examples{
\dontrun{
# Add row with generated date using standard values
gris_virus_slaktegris_utvalg <- append_date_generated_line(gris_virus_slaktegris_utvalg)
}

}
\author{
Petter Hopp Petter.Hopp@vetinst.no
}
