% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/append_sum_line.R
\name{append_sum_line}
\alias{append_sum_line}
\title{Append row with column sums}
\usage{
append_sum_line(data, column, pretext = "Sum", position = "left")
}
\arguments{
\item{data}{Data frame to which a row should be appended.}

\item{column}{Character vector. The column names of columns to sum.}

\item{pretext}{The explaining text before the sum, defaults to "Sum".}

\item{position}{The position for the pretext, on of c("first", left", "none"). defaults to left.}
}
\value{
A data frame with an appended row with sums.
}
\description{
Appends a new row with column sums for selected columns. A pretext can be placed on the row.
}
\details{
One row is appended to the data frame. The sum is calculated with na.rm = TRUE.

    If a tibble, it is transformed to a data frame to avoid errors if the pretext is to be placed in a numeric variable.
}
\examples{
\dontrun{
# Append row with sum
gris_blodprover_slakteri <- append_sum_line(data = gris_blodprover_slakteri,
                                            col_name = c("ant_prover"),
                                            pretext = "Sum",
                                            position = "first")
}

}
\author{
Petter Hopp Petter.Hopp@vetinst.no
}
