% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_random.R
\name{make_random}
\alias{make_random}
\title{Add new column with random numbers}
\usage{
make_random(data, colname = "random", seed = -1, init_seed = FALSE)
}
\arguments{
\item{data}{[\code{data.frame}]\cr
Data to which a column with random number should be added.}

\item{colname}{[\code{character(1)}]\cr
The name of the new column with the random number. Defaults to "random".}

\item{seed}{[\code{numeric(1)}]\cr
The initializing seed. Defaults to -1, see
\ifelse{html}{\code{\link[base]{set.seed}}}{\code{base::set.seed}}.}

\item{init_seed}{[\code{logical(1)}]\cr
Should the seed be initialized. Defaults to \code{FALSE}.}
}
\value{
\code{data.frame} with a new column with random numbers.
}
\description{
Adds new column with random numbers. The function is built to be
    able to use it in piping.
}
\details{
To make reproducible random numbers the seed can be initialized with a
    specific value. The first time the seed is used, set \code{init_seed = TRUE}.
    Thereafter, use \code{init_seed = FALSE} if more random numbers are generated
    in the session to avoid overlapping random numbers.
}
\examples{
\dontrun{
# Add column with random variables
x <- as.data.frame(c(1:10))
seed <- 12345

# Initialize with seed first time used
x <- make_random(x, seed = seed, init_seed = TRUE)

# Do not initialize the seed thereafter to avoid overlapping
x <- make_random(x, seed = seed, init_seed = FALSE)

# If you initialize again you get overlapping seeds
x <- make_random(x, seed = seed, init_seed = TRUE)
}

}
\author{
Petter Hopp Petter.Hopp@vetinst.no
}
