% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OKplan-deprecated.R
\name{OKplan-deprecated}
\alias{OKplan-deprecated}
\title{Deprecated Functions in Package OKplan}
\arguments{
\item{\dots}{(arguments)}
}
\value{
(results)
}
\description{
These functions are provided for compatibility with older
    versions of OKplan only, and may be defunct as soon as the next release.
    When possible, alternative functions are mentioned. Help pages for
    deprecated functions are available at \code{help("<function>-deprecated")}.
}
\details{
\code{check_OK_selection} was deprecated 2022-12-15 to replace it
    with check_ok_selection with a standardised function name with lower case
    letters for OK. Use check_ok_selection with the additional parameters
    \code{purpose =} and \code{plan_aar =}.
}
\examples{
\dontrun{
check_OK_selection(...) ### -- use \code{\link{check_ok_selection}} instead.
}
}
\author{
Petter Hopp Petter.Hopp@vetinst.no
}
\keyword{internal}
