% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OKplan-package.R
\docType{package}
\name{OKplan-package}
\alias{OKplan}
\alias{OKplan-package}
\title{OKplan: Tools to facilitate the Planning of the annual Surveillance Programmes}
\description{
Provide tools to facilitate the planning of the annual surveillance programmes. The main focus is tools for generating standardized lists for NFSA.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/NorwegianVeterinaryInstitute/OKplan}
  \item Report bugs at \url{https://github.com/NorwegianVeterinaryInstitute/OKplan/issues}
}

}
\author{
\strong{Maintainer}: Petter Hopp \email{Petter.Hopp@vetinst.no}

Other contributors:
\itemize{
  \item Norwegian Veterinary Institute [copyright holder]
}

}
\keyword{internal}
