% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_okplan.R
\name{read_okplan}
\alias{read_okplan}
\title{Reads the surveillance selection from a standardised file.}
\usage{
read_okplan(
  filename,
  filepath,
  column_standards = OKplan::OK_column_standards,
  dbsource = "okplan"
)
}
\arguments{
\item{filename}{[\code{character(1)}]\cr
The name of the csv file including extension.}

\item{filepath}{[\code{character(1)}]\cr
The path to the csv file.}

\item{column_standards}{[\code{data.frame} | \code{list} | \code{character(1)}]\cr
The column standards to be used as input for
    \ifelse{html}{\code{\link[NVIdb]{standardize_columns}}}{\code{NVIdb::standardize_columns}}
    when formatting the sampling plan for output, see details. Defaults to
    \code{\link{OK_column_standards}}.}

\item{dbsource}{[\code{character(1)}]\cr
The name of the dbtable in \code{\link{OK_column_standards}} that should
be used for standardising and formatting the sampling plan output.
Defaults to \code{"okplan"}.}
}
\value{
\code{data.frame}.
}
\description{
Reads the surveillance selection from a standardised file. The
    file has previously been saved with \code{save_okplan}. This should have
    been standardised to include a standard set of columns and have
    the records in a standard order.
}
\details{
The data is an "okplan" csv-file.
    \ifelse{html}{\code{\link[NVIdb]{standardize_columns}}}{\code{NVIdb::standardize_columns}}
    is used to standardise identify character columns which is used as input
    to \code{colClasses} in
    \ifelse{html}{\code{\link[utils]{read.csv2}}}{\code{utils::read.csv2}}.
}
\examples{
\dontrun{
library(OKplan)
td <- tempdir()
okplan <- data.frame("ok_hensiktkode" = c("01002", "01002"),
                     "ok_driftsformkode" = c("010202", "010202"),
                     "statuskode" = c(1, 1),
                     "prioritet_av_reserve" = c(NA, NA),
                     "eier_lokalitetnr" = c("1101123456", "1102123456"))
save_okplan(data = okplan,
            filename = "okplan_species_disease",
            filepath = td)
okplan <- read_okplan(filename = "okplan_species_disease",
                      filepath = td)
}
}
\author{
Petter Hopp Petter.Hopp@vetinst.no
}
