% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_okplan.R
\name{save_okplan}
\alias{save_okplan}
\title{Writes the surveillance selection to a standardised file.}
\usage{
save_okplan(
  data,
  filename,
  filepath,
  sortvar = c("ok_hensiktkode", "ok_driftsformkode", "statuskode",
    "prioritet_av_reserve", "eier_lokalitetnr"),
  ...
)
}
\arguments{
\item{data}{[\code{data.frame}]\cr
The sampling plan with the units to be reported.}

\item{filename}{[\code{character(1)}]\cr
The name of the csv file including extension.}

\item{filepath}{[\code{character(1)}]\cr
The path to the csv file.}

\item{sortvar}{[\code{character}]\cr
The sort order for the records in the csv-file. Defaults to
c("ok_hensiktkode", "ok_driftsformkode", "statuskode",
  "prioritet_av_reserve", "eier_lokalitetnr")}

\item{\dots}{Other arguments to be passed to
\ifelse{html}{\code{\link[utils:write.csv2]{utils::write.csv2}}}{\code{utils::write.csv2}}.}
}
\value{
None. Saves a data frame with the selection in a standard csv file.
}
\description{
The surveillance selection is written to a standardised file.
    The file is standardised to include a standard set of columns and have
    the records in a standard order.
}
\details{
The data is saved as an "okplan" file that will be used the source
    file when the selection list is generated.

    The function uses
    \ifelse{html}{\code{\link[NVIdb:standardize_columns]{NVIdb::standardize_columns}}}{\code{NVIdb::standardize_columns}}.
    to select and order the columns. The formatting information is taken from
    \code{\link{OK_column_standards}}.
}
\examples{
\dontrun{
library(OKplan)
td <- tempdir()
okplan <- data.frame("ok_hensiktkode" = c("01002", "01002"),
                     "ok_driftsformkode" = c("010202", "010202"),
                     "statuskode" = c(1, 1),
                     "prioritet_av_reserve" = c(NA, NA),
                     "eier_lokalitetnr" = c("1101123456", "1102123456"))
save_okplan(data = okplan,
            filename = "okplan_species_disease",
            filepath = td)
}
}
\author{
Petter Hopp Petter.Hopp@vetinst.no
}
