% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/style_sum_line.R
\name{style_sum_line}
\alias{style_sum_line}
\title{Style row with "Sum" in Excel sheet}
\usage{
style_sum_line(
  workbook = workbook,
  sheet = sheet,
  data,
  text = "Sum",
  text_decoration = "bold",
  ...
)
}
\arguments{
\item{workbook}{The workbook object.}

\item{sheet}{The Excel sheet name.}

\item{data}{The data frame that have been exported to the Excel sheet. Used to
find column number and row number for the pretext for which the row should be styled.}

\item{text}{The text in the cell for which the row should be styled.
Defaults to "Sum".}

\item{text_decoration}{The text decoration style that should be used, see \code{openxlsx::createStyle}.
Defaults to "bold".}

\item{\dots}{Other arguments to be passed.}
}
\value{
None. One row in the workbook object is styled.
}
\description{
Style the font with bold for the line with the text "Sum" in on cell.
    It is possible to use other text decoration, see \code{openxlxs::createStyle}.
    A line with the text "Sum" or another text as given by text will be styled.
}
\details{
The whole line will be styled.
}
\author{
Petter Hopp Petter.Hopp@vetinst.no
}
