% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/append_date_generated_line.R
\name{append_date_generated_line}
\alias{append_date_generated_line}
\title{Add new last row with generated date}
\usage{
append_date_generated_line(
  data,
  pretext = "Datauttrekket er gjort",
  date = format(Sys.Date(), "\%d/\%m/\%Y")
)
}
\arguments{
\item{data}{Data frame with data that should get new row with generated date}

\item{pretext}{The explaining text before the date value.}

\item{date}{Date for generating the data}
}
\value{
A data frame with two more rows, one empty and one with generated data in the first column.
}
\description{
Function to add new row with generated date. Before the date the pretext is "Datauttrekket ble gjort"
}
\details{
Two rows are dded to the data frame, the first is empty, the second has the generated date in the first column. The
}
\examples{
\dontrun{
# Add row with generated date using standard values
gris_virus_slaktegris_utvalg <- append_date_generated_line(gris_virus_slaktegris_utvalg)

}

}
\author{
Petter Hopp Petter.Hopp@vetinst.no
}
