% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/append_sum_line.R
\name{append_sum_line}
\alias{append_sum_line}
\title{Append row with column sums}
\usage{
append_sum_line(data, column, pretext = "Sum", position = "left")
}
\arguments{
\item{data}{[\code{data.frame}]\cr
Data to which a row should be appended.}

\item{column}{[\code{character}]\cr
The column names of columns to sum.}

\item{pretext}{[\code{character(1)}]\cr
The explaining text before the sum. Defaults to "Sum".}

\item{position}{[\code{character(1)}]\cr
The position for the pretext, one of c("first", left", "none"). Defaults
to "left".}
}
\value{
\code{data.frame} with an appended row with sums.
}
\description{
Appends a new row with column sums for selected columns. A
    pretext can be placed on the row.
}
\details{
One row is appended to the data frame. The sum is calculated with
    \code{na.rm = TRUE}.

    If a tibble, it is transformed to a data frame to avoid errors if the
    pretext is to be placed in a numeric variable.
}
\examples{
\dontrun{
# Append row with sum
gris_blodprover_slakteri <- append_sum_line(data = gris_blodprover_slakteri,
                                            column = c("ant_prover"),
                                            pretext = "Sum",
                                            position = "first")
}

}
\author{
Petter Hopp Petter.Hopp@vetinst.no
}
