% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_OK_selection-deprecated.R
\name{check_OK_selection-deprecated}
\alias{check_OK_selection-deprecated}
\title{Control of standard output file with OK selection}
\usage{
check_OK_selection(data)
}
\arguments{
\item{data}{Data frame with selection for a OK programme.}
}
\value{
Prints results of the control to the output window.
}
\description{
Standard control by performing descriptive statistics of variables in the output file with OK selection.
}
\details{
Under development. This should be rewritten to produce nicer output.

    Gives descriptive statistics of the selection. This is used to see if the number of selected units per category are correct.
    If any mistakes are found, one must go back and correct in the script that produces the selection.

    Before the control is run, the column names must have been standardized using \code{NVIdb::standardize_columns}.
}
\examples{
\dontrun{
# Checking OK selection data in data frame okplan_MRSA
check_OK_selection(okplan_MRSA)
}

}
\author{
Petter Hopp Petter.Hopp@vetinst.no
}
\keyword{internal}
