% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_OK_selection-deprecated.R
\name{check_OK_selection}
\alias{check_OK_selection}
\title{check_OK_selection is Deprecated}
\usage{
check_OK_selection(
  data,
  purpose = deparse(substitute(data)),
  plan_aar = as.numeric(format(Sys.Date(), "\%Y")) + 1
)
}
\arguments{
\item{data}{The table with data describing the selection for a OK programme.}

\item{purpose}{String with descriptive text to be used in file name and
heading of the report. Defaults to name of input data.}

\item{plan_aar}{The year for which the
selection is planned. Defaults to next year.}
}
\description{
\code{check_OK_selection} was deprecated 2022-12-15 to replace it
    with check_ok_selection with a standardised function name with lower case
    letters for OK. Use check_ok_selection with the additional parameters
    \code{purpose =} and \code{plan_aar =}.
}
\details{
The old help pages can be found at \code{help("check_OK_selection-deprecated")}.
    Information on deprecated function can be found at \code{help("OKplan-deprecated")}.
}
