% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_tested_herds.R
\name{get_tested_herds}
\alias{get_tested_herds}
\title{Gets herds tested within an surveillance programme}
\usage{
get_tested_herds(
  eos_table,
  year = as.numeric(format(Sys.Date(), "\%Y")) - 1,
  species = NULL,
  production = NULL,
  disease = NULL,
  min_prover = -1,
  tested = FALSE
)
}
\arguments{
\item{eos_table}{[\code{character(1)}]\cr
EOS table name.}

\item{year}{[\code{numeric(1)}]\cr
One or more years that should be selected. Defaults
to previous year.}

\item{species}{[\code{character}]\cr
The species that should be selected. Defaults to \code{NULL}.}

\item{production}{[\code{character}]\cr
The production type(s) for which number of tested samples should be
calculated. Defaults to \code{NULL}.}

\item{disease}{[\code{character(1)}]\cr
The disease for which number of tested samples should be calculated.
Defaults to \code{NULL}.}

\item{min_prover}{[\code{numeric(1)}]\cr
Minimum number of samples that should have been received or examined for
the herd to be counted as sampled or tested. No check is performed if
equal -1. Defaults to -1.}

\item{tested}{[\code{logical(1)}]\cr
If \code{TRUE}, the number of tested samples, If \code{FALSE}, the number
of received samples. Defaults to \code{FALSE}.}
}
\value{
\code{data.frame} with tested or sampled locations.
}
\description{
Gets herds that have been sampled or tested within
    a surveillance programme for the selected years. You can
    choose between herds that have submitted samples or herds
    for which a certain number of samples have been examined
    for a specific disease.
}
\details{
For programmes having several surveillance streams, it is possible
    to select surveillance streams based on species and production type. The
    species and/or production type must be written as in the eos-table. Be
    aware that species and production type may be missing and for production
    type it may often be wrong. Therefore, selection by production type and/or
    species may remove saker that you would want to keep. No selection is
    performed when the species or production type is missing from the
    eos-table.

It is possible to define a minimum requirement of number of
    samples received or tested. For programmes covering several
    infections, it is necessary to input the disease for which
    the samples should have been tested. The disease name must
    be given as it is written in the column name for the number
    of examined samples.

The eos_table name is the same name as the table name in the EOS data base.
}
\author{
Petter Hopp Petter.Hopp@vetinst.no
}
