% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_ok_selection_list.R
\name{write_ok_selection_list}
\alias{write_ok_selection_list}
\title{Writes an Excel file with the selection list.}
\usage{
write_ok_selection_list(
  data,
  filename = paste0(NVIdb::set_dir_NVI("OKprogrammer"), "Rutine", plan_aar,
    "/planlegging/Utvalgslister/", "Storfe ammeku BVD EBL IBR ", plan_aar,
    " Antall blodprøver per slakteri.xlsx"),
  sheet = paste0("Antall_blodprøver_ammeku_", plan_aar),
  dbtemplate
)
}
\arguments{
\item{data}{The data with units that should be tested.}

\item{filename}{The name and path of the Excel file that should be written.}

\item{sheet}{The name of the Excel sheet with the list.}

\item{dbtemplate}{The name of the dbtable in OK_column_standards that should
be used for standardizing the columns.}
}
\description{
The selection list based on selected data from okplan file and
    uses standardize_columns to  select, format and style columns.
}
