% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_ok_selection_list.R
\name{write_ok_selection_list}
\alias{write_ok_selection_list}
\title{Writes an Excel file with the selection list.}
\usage{
write_ok_selection_list(
  data,
  sheet,
  filename,
  filepath,
  column_standards = OKplan::OK_column_standards,
  calculate_sum = TRUE,
  footnote = NULL,
  footnote_heights = NULL,
  dbsource,
  add_worksheet = FALSE
)
}
\arguments{
\item{data}{[data.frame]. The data with the selected units that should be reported.}

\item{sheet}{[character(1)]. The name of the Excel sheet with the list.}

\item{filename}{[character(1)]. The name of the Excel file that should be written.}

\item{filepath}{[character(1)]. The path to the Excel file that should be written.}

\item{column_standards}{[data.frame | list]. The column standards for OK data frames to be used as
input for \code{NVIdb::standardize_columns} when generating the selection
list, see details. Defaults to \code{OKplan::OK_column_standards}.}

\item{calculate_sum}{[logical(1)]. Should a line with the sum be
appended? Defaults to \code{TRUE}.}

\item{footnote}{[character(1)]. Footnote to appended? Defaults to \code{NULL}.}

\item{footnote_heights}{[integer(1)]. Manually set row height for the
footnote. Defaults to \code{NULL}.}

\item{dbsource}{[character(1)]. The name of the dbtable in OK_column_standards that should
be used for standardizing the columns.}

\item{add_worksheet}{[logical(1)]. Should a worksheet be added to
an existing workbook? Defaults to \code{TRUE}.}
}
\description{
The selection list generated and ready formatted to
be sent.
}
\details{
The data must originate from an "okplan" file and
the function uses standardize_columns to  select, order, format and
style the columns. The formatting information is either taken
from \code{OKplan::OK_column_standards} or can be input as
a list.

When using \code{OKplan::OK_column_standards}, the formatting
information is taken in accord with the argument \code{dbsource}.
If the formatting needs to be edited, it must be edited in the
general source file for column standards and incorporated into
\code{OKplan}. As this can be a tedious process, there is a
possibility to input the formatting information as a list.

The list input to column_standards must follow a specific format.
It is a list with at least three vectors:
\itemize{
\item colname = a vector of all columns in in the source file that
should be included in the Excel report with the selection list.
\item collabel: A vector with the labels that should be used in the
Excel report.
\item colwidth: A vector with the column width that should be used
in the Excel report.
}

All vectors must have the same order and the same length.
In addition one may input:
\itemize{
\item colorder: the order of the columns to be used in the Excel report.
The Default is to use the same order as they are entered in the vectors.
\item table_db: input added as a possibility to keep on the same format
as \code{OKplan::OK_column_standards}. Not necessary to input.
\item column_db: input added as a possibility to keep on the same format
as \code{OKplan::OK_column_standards}. Must be the same as
\code{dbsource}. Not necessary to input.
}

When more than one worksheet should be added to a single workbook,
use \code{add_worksheet = FALSE} for the first worksheet and
\code{add_worksheet = TRUE} for the consecutive worksheet(s).
}
