% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_ok_selection_list.R
\name{write_ok_selection_list}
\alias{write_ok_selection_list}
\title{Writes the sampling plan with the selection list to an Excel file.}
\usage{
write_ok_selection_list(
  data,
  sheet,
  filename,
  filepath,
  column_standards = OKplan::OK_column_standards,
  calculate_sum = TRUE,
  footnote = NULL,
  footnote_heights = NULL,
  dbsource,
  add_worksheet = FALSE,
  ...
)
}
\arguments{
\item{data}{[\code{data.frame}]\cr
The sampling plan with the units to be reported.}

\item{sheet}{[\code{character(1)}]\cr
The name of the Excel sheet.}

\item{filename}{[\code{character(1)}]\cr
The name of the Excel file.}

\item{filepath}{[\code{character(1)}]\cr
The path to the Excel file.}

\item{column_standards}{[\code{data.frame} | \code{list}]\cr
The column standards to be used as input for
\ifelse{html}{\code{\link[NVIdb]{standardize_columns}}}{\code{NVIdb::standardize_columns}}
when formatting the sampling plan for
output, see details. Defaults to \code{\link{OK_column_standards}}.}

\item{calculate_sum}{[\code{logical(1)}]\cr
Should a line with the sum be appended? Defaults to \code{TRUE}.}

\item{footnote}{[\code{character(1)}]\cr
Footnote to appended? Defaults to \code{NULL}.}

\item{footnote_heights}{[\code{integer(1)}]\cr
Manually set row height for the footnote. Defaults to \code{NULL}.}

\item{dbsource}{[\code{character(1)}]\cr
The name of the dbtable in \code{\link{OK_column_standards}} that should
be used for standardising and formatting the sampling plan output.}

\item{add_worksheet}{[\code{logical(1)}]\cr
Should a worksheet be added to an existing workbook? Defaults to
\code{FALSE}.}

\item{\dots}{Other arguments to be passed to \code{\link{append_sum_line}}.}
}
\description{
The sampling plan is output to an Excel sheet. The list with
    selected units is standardised and formatted in order to be submitted
    without further formatting.
}
\details{
The data must originate from an "okplan" file and
    the function uses standardize_columns to  select, order, format and
    style the columns. The formatting information is either taken
    from \code{\link{OK_column_standards}} or can be input as
    a list.

When using \code{\link{OK_column_standards}}, the formatting information is
    taken in accord with the argument \code{dbsource}. If the formatting
    needs to be edited, it must be edited in the general source file for
    column standards and thereafter, built into a new version of \code{OKplan}.
    As this can be a tedious process, there is a possibility to input the
    formatting information as a list.

The list input to column_standards must follow a specific format.
    It is a list with at least three named vectors:
\itemize{
\item \code{colname}: a vector of all columns in in the source file that
    should be included in the Excel report with the selection list.
\item \code{collabel}: A vector with the labels that should be used in the
    Excel report.
\item \code{colwidth}: A vector with the column width that should be used
    in the Excel report.
}

    In addition one may input:

\itemize{
\item \code{colorder}: the order of the columns to be used in the Excel report.
    The default is to use the same order as they are entered in the vectors.
\item \code{column_db}: input added as a possibility to keep on the same format
    as \code{\link{OK_column_standards}}. Not necessary to input.
\item \code{table_db}: input added as a possibility to keep on the same format
    as \code{\link{OK_column_standards}}. Must be the same as
    \code{dbsource}. Not necessary to input.
}

All vectors must have the same order and the same length.

When more than one worksheet should be added to a single workbook,
    use \code{add_worksheet = FALSE} for the first worksheet and
    \code{add_worksheet = TRUE} for the consecutive worksheet(s).
}
