% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_ok_selection_list.R
\name{write_ok_selection_list}
\alias{write_ok_selection_list}
\title{Writes an Excel file with the selection list.}
\usage{
write_ok_selection_list(
  data,
  sheet,
  filename,
  filepath,
  calculate_sum = TRUE,
  dbsource
)
}
\arguments{
\item{data}{The data with units that should be tested.}

\item{sheet}{The name of the Excel sheet with the list.}

\item{filename}{The name of the Excel file that should be written.}

\item{filepath}{The path to the Excel file that should be written.}

\item{calculate_sum}{\[logical\] Should a line with the sum be appended. Defaults to TRUE.}

\item{dbsource}{The name of the dbtable in OK_column_standards that should
be used for standardizing the columns.}
}
\description{
The selection list based on selected data from okplan file and
    uses standardize_columns to  select, format and style columns.
}
